Add javascript date post

+++ _gulp/blog.js
@@ -12,8 +12,10 @@ var blogPath = sitePath + '/blog';
 
 module.exports = function(done) {
 
-  // move blog index to root dir
-  fs.renameSync(blogPath + '/index.html', sitePath + '/blog.html');
+  if (fs.existsSync(blogPath + '/index.html')) {
+    // move blog index to root dir
+    fs.renameSync(blogPath + '/index.html', sitePath + '/blog.html');
+  }
 
   // traverse every blog index file
   fsTools.walkSync(blogPath, 'index.html$', function(path) {
@@ -22,11 +24,14 @@ module.exports = function(done) {
     var dir = /\/([^\/]+)\/index\.html$/.exec(path)[1];
     var newPath = blogPath + '/' + dir + '.html';
 
-    // move to blog dir
-    fs.renameSync(path, newPath);
+    if (fs.existsSync(path)) {
 
-    // remove folder
-    fs.rmdirSync(blogPath + '/' + dir);
+      // move to blog dir
+      fs.renameSync(path, newPath);
+
+      // remove folder
+      fs.rmdirSync(blogPath + '/' + dir);
+    }
 
   });
